+++ pkg/generator/generator.go
@@ -316,7 +316,15 @@ func applyPatch(ctx context.Context, base, patch []byte) (string, error) {
 		c := pathmodifier.DefaultConfig()
 		c.InputBytes = base
 		svc, err := pathmodifier.New(c)
-		if err != nil {
+		if pathmodifier.IsInvalidFormat(err) {
+			// return a descriptive error message
+			x := struct{}{}
+			yamlErr := yaml.Unmarshal([]byte(base), &x)
+			if yamlErr != nil {
+				return "", microerror.Mask(yamlErr)
+			}
+			return "", microerror.Mask(err)
+		} else if err != nil {
 			return "", microerror.Mask(err)
 		}
 		basePathSvc = svc
@@ -327,7 +335,15 @@ func applyPatch(ctx context.Context, base, patch []byte) (string, error) {
 		c := pathmodifier.DefaultConfig()
 		c.InputBytes = patch
 		svc, err := pathmodifier.New(c)
-		if err != nil {
+		if pathmodifier.IsInvalidFormat(err) {
+			// return a descriptive error message
+			x := struct{}{}
+			yamlErr := yaml.Unmarshal([]byte(patch), &x)
+			if yamlErr != nil {
+				return "", microerror.Mask(yamlErr)
+			}
+			return "", microerror.Mask(err)
+		} else if err != nil {
 			return "", microerror.Mask(err)
 		}
 		patchPathSvc = svc