+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Return a more descriptive error when given invalid YAML.
+
 ## [0.2.9] - 2021-04-02
 
 ### Fixed